pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [
    { include = 'gendiff' },
]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.1.3"
flake8 = "^5.0.4"
pytest-cov = "^4.0.0"
coverage = "^6.5.0"
PyYAML = "^6.0"
yamllint = "^1.28.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = 'gendiff.scripts.gendiff_script:main'